test: expect padding in fullscreen view box
